Rename yagna_dapp_manager to dapp_manager
