+++ dapp_manager/__init__.py
@@ -1,0 +1,2 @@
+from dapp_manager.dapp_manager import DappManager
+from dapp_manager.exceptions import UnknownApp
