+++ dapp_manager/autocomplete/__init__.py
@@ -6,7 +6,7 @@ import click
 # Entrypoint aka binary name aka script name
 # Entrypoint must match the name defined in pyproject.toml under
 # [tool.poetry.scripts] section
-ENTRYPOINT = "yagna_dapp_manager"
+ENTRYPOINT = "dapp_manager"
 COMPLETION_FUNC_NAME = f"{ENTRYPOINT}_completion"
 SCRIPTS_DIR = Path(__file__).parent / "scripts"
