+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "yagna-dapp-manager"
+name = "dapp-manager"
 version = "0.0.1"
 description = "Golem dapp-manager - run decetralized apps on the Golem Network using a 'docker-compose'-like interface"
 authors = ["Golem Factory <contact@golem.network>"]
@@ -32,7 +32,7 @@ pytest = "^7.1"
 pytest-cov = "^3.0.0"
 
 [tool.poetry.scripts]
-yagna_dapp_manager = "yagna_dapp_manager.__main__:main"
+dapp_manager = "dapp_manager.__main__:main"
 
 [tool.poe.tasks]
 codestyle = "black --check --diff ."
