--- yagna_dapp_manager/__init__.py
@@ -1,2 +1,0 @@
-from yagna_dapp_manager.dapp_manager import DappManager
-from yagna_dapp_manager.exceptions import UnknownApp
